% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{BlockAverageMod}
\alias{BlockAverageMod}
\title{Block average a time series with either consecutive or non-consecutive blocks}
\usage{
BlockAverageMod(dts, tsin, blockwidth = NULL)
}
\arguments{
\item{dts}{Middle points of the time steps for the block averaged process}

\item{tsin}{Higher resolution process to be block averaged, should be a zoo object}

\item{blockwidth}{block width for the sampling points dts. Defaults to NULL (consecutive blocks). If a numeric value is given, this is taken as the width (in dts units) around the dts-midpoints, if a vector (of the same length as dts) is given, these are the widthes around dts for the sampling}
}
\value{
\code{zoo}-object with the block-averaged time series
}
\description{
Block average a time series with either consecutive or non-consecutive blocks
}
\examples{
dts<-seq(5,1050,by=55)
tsin<-zoo(rnorm(1000),order.by=seq(1,1000,by=1))

ts.noncons<-BlockAverageMod(dts=dts,tsin=tsin,blockwidth=20) # an error occurs if dtmin is smaller than blockwidth/2
ts.cons<-BlockAverageMod(dts=dts,tsin=tsin)
ts.noncons.changing<-BlockAverageMod(dts=dts,tsin=tsin,blockwidth=seq(5,50,length.out=length(dts)))
plot(tsin)
lines(ts.noncons,col="blue",lwd=2)
lines(ts.cons,col="cyan")
lines(ts.noncons.changing,col="green")
}
\author{
Kira Rehfeld (added nonconsecutive option) & Raphaël Hébert (consecutive original)
}
