% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check.Dtable}
\alias{check.Dtable}
\title{check a dating table and classify the difficulty and treatment level}
\usage{
check.Dtable(speleo, add = TRUE)
}
\arguments{
\item{speleo}{\code{speleo}-object}

\item{add}{\code{logical} indicating plotting or not}
}
\value{
\code{integer}, 0 for trivial (monotonous), 1 for slight reversals, 2 for required treatment
}
\description{
check a dating table and classify the difficulty and treatment level
}
