% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{chk.which.rev}
\alias{chk.which.rev}
\title{Check which ages are reversed, and which one should be removed}
\usage{
chk.which.rev(AgeEst, DepthS, UncEst = rep(0, length(AgeEst)))
}
\arguments{
\item{AgeEst}{vector of ages}

\item{DepthS}{vector of depths}

\item{UncEst}{vector of age uncertainties (1 sd)}
}
\value{
index
}
\description{
Check which ages are reversed, and which one should be removed
}
