% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do.dating}
\alias{do.dating}
\title{do the dating of the synthetic speleothem}
\usage{
do.dating(speleo)
}
\arguments{
\item{speleo}{speleo-object, containing \code{datingparams} list with \code{nsample}, the number of point-ages sampled, \code{precision}, the sample age uncertainty for t-distribution (width equivalent to standard deviation if \code{tdofs}=\code{infinity}), and \code{tdofs}, the degrees of freedom for the t-distribution}
}
\value{
speleo speleo-object with updated \code{Dtable} and \code{datingparams}
}
\description{
Simulate a dating sample on a synthetic speleothem
}
\seealso{
\code{\link{generate.speleo}} \code{\link{plotgrowthfcn}} \code{\link{do.dating}}
}
