% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{export.files}
\alias{export.files}
\title{Write the files for the synthetic speleothem record to disk}
\usage{
export.files(speleo, writepath = "./")
}
\arguments{
\item{speleo}{\code{speleo}-object}

\item{writepath}{valid path}
}
\description{
Write the files for the synthetic speleothem record to disk
}
