% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate.speleo}
\alias{generate.speleo}
\title{Generate synthetic speleothem record}
\usage{
generate.speleo(t.res = 1, z.top = 0, z.bottom = 1000, t.top = 0,
  t.bottom = 800, growthfcn = NA, Dtable = NA, proxy = NA,
  name = "Synthetic")
}
\arguments{
\item{t.res}{temporal resolution of the initial simulation}

\item{z.top}{top depth (assumed: in millimeters)}

\item{z.bottom}{bottom (assumed: in millimeters)}

\item{t.top}{top age (assumed: in years before present)}

\item{t.bottom}{bottom age (assumed: in years before present)}

\item{growthfcn}{growth object (dummy parameter)}

\item{Dtable}{dating table}

\item{proxy}{proxy object (dummy parameter)}

\item{name}{Name of the speleothem}
}
\value{
speleo-object (list)
}
\description{
Generate a synthetic speleothem record
}
\examples{
speleo<-generate.speleo()
speleo<-growthfcn.constantgamma(speleo)
plotgrowthfcn(speleo)
speleo<-do.dating(speleo)
plotdating(speleo$Dtable,add=TRUE)
export.files(speleo)
}
\seealso{
\code{\link{generate.speleo}} \code{\link{plotgrowthfcn}} \code{\link{do.dating}}
}
