% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get.simpleagemodel}
\alias{get.simpleagemodel}
\title{Get simple age model}
\usage{
get.simpleagemodel(speleo, itype = "pchip", nrealiz = 10, add = FALSE,
  dtype = "tdist", nfac = nrealiz * 10)
}
\arguments{
\item{speleo}{\code{speleo}-object}

\item{itype}{interpolation type "linint" or "spline"}

\item{nrealiz}{\code{integer}-value}

\item{add}{\code{logical}, default \code{FALSE}, if true plots results}

\item{dtype}{date type "tdist" or "norm" for "t-distributed" or "gaussian-distributed" date sampling}

\item{nfac}{\code{integer} value, should be larger than \code{nrealiz}, much larger if there are reversals}
}
\description{
Get simple age model
}
