% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{growthfcn.constantgamma}
\alias{growthfcn.constantgamma}
\alias{growthfcn.changinggamma}
\title{Gamma-distributed growth function with constant mean growth}
\usage{
growthfcn.constantgamma(speleo, skew = 1)

growthfcn.changinggamma(speleo, tlist = c(speleo$t.top, (speleo$t.bottom -
  speleo$t.top)/4, (speleo$t.bottom - speleo$t.top)/2, 3 * (speleo$t.bottom -
  speleo$t.top)/4, speleo$t.bottom), rates = c(1, 0, 2, 4), skew = 1)
}
\arguments{
\item{speleo}{\code{speleo}-object}

\item{skew}{skewness of the growth distribution}

\item{tlist}{vector of start and end times for growth rate changes}

\item{rates}{slopes (must be of length \code{L=length(tlist)-1})}
}
\value{
speleo-object \code{speleo}-object with updated \code{growthfcn}
}
\description{
Simulate the growth of a constantly accumulating (speleothem) palaeoclimate record.
}
\examples{
speleo<-generate.speleo()
growthfcn<-growthfcn.constantgamma(speleo)
plotgrowthfcn(growthfcn)

speleo<-generate.speleo()
growthfcn<-growthfcn.changinggamma(speleo)
plotgrowthfcn(growthfcn)
}
\seealso{
\code{\link{generate.speleo}} \code{\link{plotgrowthfcn}} \code{\link{do.dating}}
}
