% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{kim}
\alias{kim}
\title{Kim & o'Neill temperature dependent calcite fractionation}
\usage{
kim(temperature = NULL, alpha = NULL, delta = NULL, tempdelta = NULL)
}
\arguments{
\item{temperature}{temperature in kelvin}

\item{alpha}{fractionation factor R(A)/R(B)}

\item{delta}{difference sea water/calcite d18O in permil (Ganssen et al. 2011, referencing to Kim & o'Neil 1997)}

\item{tempdelta}{inverted delta}
}
\value{
numeric value (if \code{temperature} is given, return \code{alpha}, if \code{alpha} is given, return \code{temperature}; if \code{delta} is given, return \code{temperature} )
}
\description{
Kim & o'Neill temperature dependent calcite fractionation
}
\examples{
kim(alpha=kim(temperature=283.15))
curve(kim,from=273.15,to=300,xlab="temperature",ylab="alpha")
kim(delta=2)
# --> The Glacial-Interglacial change of approx. 2 permil would be equivalent to approximately 7K.
}
\references{
Ganssen et al., (2011), CLim. Past http://dx.doi.org/10.5194/cp-7-1337-2011

Kim, S.-T., & O’Neil, J. R. (1997). Equilibrium and nonequilibrium oxygen isotope effects in synthetic carbonates. Geochimica et Cosmochimica Acta, 61(16), 3461–3475. https://doi.org/10.1016/S0016-7037(97)00169-5
}
