% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotdating}
\alias{plotdating}
\title{Plot the dating on top of the growth function of the synthetic speleothem}
\usage{
plotdating(Dtable, add = FALSE)
}
\arguments{
\item{Dtable}{from \code{speleo}-object and/or growth function}

\item{add}{logical, if TRUE plot.}
}
\description{
Plot the dating table information on the growth function
}
\examples{
speleo<-generate.speleo()
speleo<-growthfcn.constantgamma(speleo)
plotgrowthfcn(speleo)
speleo<-do.dating(speleo)
plotdating(speleo$Dtable,add=TRUE)
export.files(speleo)
}
\seealso{
\code{\link{generate.speleo}} \code{\link{plotgrowthfcn}} \code{\link{do.dating}}
}
