% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotgrowthfcn}
\alias{plotgrowthfcn}
\title{plot growth function}
\usage{
plotgrowthfcn(speleo)
}
\arguments{
\item{speleo}{\code{speleo}-object}
}
\value{
\code{speleo}-object
}
\description{
Plot the growth function of the synthetic speleothem record
}
\details{
Plot the growth function of the synthetic speleothem
}
