% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample.proxy}
\alias{sample.proxy}
\title{Sample a proxy time series}
\usage{
sample.proxy(speleo, nsamp = 50, z.steps = NULL, stype = "gaussian",
  prx.unc = 1, ...)
}
\arguments{
\item{speleo}{\code{speleo}-object}

\item{nsamp}{\code{integer}}

\item{z.steps}{type of sampling, following \code{\link{BlockAverageMod}}. If \code{NULL}, consecutive sampling is expected. If equal to single number, this is the depth integrated in one sample. If a vector is given, that is expected to give the depths integrated in each sample individually.}

\item{stype}{\code{character}, currently only "gaussian" or "sinusoid" (10 cycles), if you supply a vector of length L=\code{length(speleo$true.age)} it is used as signal}

\item{prx.unc}{\code{numeric} proxy uncertainty}

\item{...}{optional}
}
\value{
\code{speleo}-object
}
\description{
Sample a proxy time series
}
