% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{TDist_ls}
\alias{TDist_ls}
\alias{dt_ls}
\alias{pt_ls}
\alias{pt_ls}
\alias{qt_ls}
\alias{rt_ls}
\title{Location-Scale formulation of the Student t distribution}
\usage{
dt_ls(x, df, mu, a)

pt_ls(x, df, mu, a)

qt_ls(prob, df, mu, a)

rt_ls(n, df, mu, a)
}
\arguments{
\item{x}{quantiles}

\item{df}{degrees of freedom}

\item{mu}{mean (location)}

\item{a}{scale (standard deviation for df=infinity)}

\item{prob}{vector of quantiles}

\item{n}{integer}
}
\value{
Analogous to the \code{\link{TDist}} implementation, \code{dt_ls} gives the density, \code{pt_ls} the distribution function, \code{qt_ls} the quantile function and \code{rt_ls} the random deviations

\code{pt_ls()} returns probabilities

\code{qt_ls()} returns quantiles

\code{rt_ls()} returns observations
}
\description{
Location-Scale formulation of the student t-distribution
}
\examples{
curve(exp=dnorm(x,mean=1,sd=3),from=-15,to=15,col="red",xlab="x",ylab="density")
curve(expr =dt_ls(x,a=3,mu=1,df=3),from=-15,to=15,add=TRUE)
curve(expr =dt_ls(x,a=3,mu=1,df=1),from=-15,to=15,add=TRUE,lty=2)
curve(expr =dt_ls(x,a=5,mu=1,df=3),from=-15,to=15,add=TRUE,lty=3)
curve(expr =dt_ls(x,a=3,mu=2,df=3),from=-15,to=15,add=TRUE,lty=4)
curve(expr=dt_ls(x,a=3,mu=2,df=30),add=TRUE,col="orange",lty=1)
legend("topleft",lty=seq(1,4),c("dt_ls:a=3,mu=1,df=3","dt_ls:a=3,mu=1,df=1","dt_ls:a=5,mu=1,df=3","dt_ls:a=3,mu=2,df=3","dnorm:mu=1,sd=3","dt_ls:a=3,mu=2,df=30"),col=c(1,1,1,1,"red","orange"))
}
\references{
\url{https://en.wikipedia.org/wiki/Location-scale_family}
}
