% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update.Dtable}
\alias{update.Dtable}
\title{Update a dating table}
\usage{
\method{update}{Dtable}(speleo, ignore.dates = NULL, increased.unc = NULL,
  date.tdofs = NULL, add = FALSE)
}
\arguments{
\item{speleo}{\code{speleo}-object}

\item{ignore.dates}{\code{vector} of dates. Defaults to \code{NULL}, or a vector of \code{NA}-values. Dates to be ignored are to be set to not-\code{NA}}

\item{increased.unc}{\code{vector} of multipliers for the dating uncertainty. Set to ones or \code{NULL} for no change.}

\item{date.tdofs}{\code{vector} of degrees of freedom for the Monte-Carlo point age sampling. Set to large values (e.g. 100) for very gaussian, small values (e.g. 1) for very skewed sampling following a t-Distribution}

\item{add}{\code{logical} describing whether to plot or not.}
}
\value{
speleo-object with updated DtableMod
}
\description{
Update a dating table
}
